+++ ibis/__init__.py
@@ -6,7 +6,7 @@ import ibis.expr.api as api  # noqa: F401
 import ibis.expr.types as ir  # noqa: F401
 
 # pandas backend is mandatory
-import ibis.pandas.api as pandas  # noqa: F401
+from ibis import pandas  # noqa: F401
 import ibis.util as util  # noqa: F401
 from ibis.common.exceptions import IbisError  # noqa: F401
 from ibis.config import options  # noqa: F401
