+++ ibis/file/csv.py
@@ -7,7 +7,7 @@ import ibis.expr.schema as sch
 from ibis.expr.scope import Scope
 from ibis.expr.typing import TimeContext
 from ibis.file.client import FileClient
-from ibis.pandas.api import PandasDialect
+from ibis.pandas import PandasDialect
 from ibis.pandas.core import execute, execute_node, pre_execute
 from ibis.pandas.execution.selection import physical_tables
