+++ ibis/pandas/execution/generic.py
@@ -25,6 +25,7 @@ from ibis.expr.scope import Scope
 from ibis.expr.timecontext import TIME_COL
 from ibis.expr.typing import TimeContext
 from .. import aggcontext as agg_ctx
+from ..client import PandasClient, PandasTable
 from ..core import (
     boolean_types,
     execute,
@@ -910,9 +911,7 @@ def execute_node_where_scalar_scalar_series(op, cond, true, false, **kwargs):
     return pd.Series(np.repeat(true, len(false))) if cond else false
 
 
-@execute_node.register(
-    ibis.pandas.client.PandasTable, ibis.pandas.client.PandasClient
-)
+@execute_node.register(PandasTable, PandasClient)
 def execute_database_table_client(
     op, client, timecontext: Optional[TimeContext], **kwargs
 ):
